refactor(backend): 重构后端结构与模型处理
测试用例详见Apifox

close #3
