+++ backend/.gitignore
@@ -1,3 +1,6 @@
+# upload files
+upload/
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
