+++ backend/README.md
@@ -11,6 +11,5 @@ pip install flask
 # optional: open debug mode
 export FLASK_ENV=development
 # run the backend server at localhost:5000
-# before that you need to build the frontend
 flask run
 ```
