+++ backend/app/utils/utils.py
@@ -1,0 +1,2 @@
+def fileExtension(fileName):
+    return fileName.rsplit('.', 1)[1].lower()
